Remove concept of base/child meshes.
Originally, only the base mesh and its children were going to be exported.
I've decided to export all valid visible meshes in the scene root if the user decides to do so.
You can make the exporter export only the active object and its child meshes if the active object is visible, and in the scene root.

Also, prepare to implement flat colour material export.