Fix crashes that occur with detail-x objects.
Fix crashes that occur when multiple detail-x objects are within a scene.
Remove LOD naming scheme stuff. Now I'm using a smarter prefix-lod-suffix system.
Fix bug that forces the root export filename to be the name of the root object.
Convert base_lod to a field of ModelManager.
Fix bug that does not allow users to use a plain object (no LOD suffixes) for a LOD 0 base/root object.