Make the model importer and MAT reader search ../mat and ../pal.
Add a new look_for method to MaterialManager and MATReader that searches case-insensitively for a file inside a directory inside the parent directory.
Remove mat_path, and make mat_fname an absolute path.
Don't create a black placeholder image for images that cannot be found. Instead, set self.mtimages[texnum] to None.
Handle the case where the NAME chunk is used (external palette) instead of the CMAP chunk in MATReader.
Generate a grayscale placeholder palette if a palette cannot be found.